refactor shared v3/v4 logic
